[MLIR] Fix casting and avoid code duplication

+++ src/vpux_compiler/include/vpux/compiler/dialect/VPU/nce_sparsity.hpp
@@ -49,7 +49,7 @@ std::vector<int32_t> getWeightsTable(mlir::Type inElemType, mlir::Type outElemTy
                                      int64_t OC, mlir::Type weightsElemType = nullptr,
                                      Const::ContentAttr bias = nullptr, VPU::PPETaskAttr ppe = nullptr);
 
-void computeQuantMultShift(double scale, int64_t& shift, int64_t& mult, uint32_t bits = 15);
+void computeQuantMultShift(double scale, uint32_t& shift, uint32_t& mult, uint32_t bits = 15);
 
 }  // namespace NCESparsity
